Optimize the integration of DepthAnything
