+++ requirements.txt
@@ -33,11 +33,11 @@ sentencepiece
 pydantic
 scikit-image
 huggingface_hub[cli]
+clip @ git+https://github.com/openai/CLIP.git
 dominate
 pycuda
 flask
 flask-socketio
-torchao
 gradio==4.14.0
 gradio_imageslider
