+++ test/test_depth_anything.py
@@ -2,23 +2,22 @@ import sys
 from pathlib import Path
 sys.path.append("..")
 
-from src.sceneflow.pipelines.depthanything.pipeline_depthanything import (  
+from src.sceneflow.pipelines.depth_anything.pipeline_depth_anything import (  
     DepthAnythingPipeline,
 )
 
 # Configure before running
 DATA_TYPE = "image"  # or "video"
 DATA_PATH = "/YOUR/IMAGE/OR/VIDEO/PATH"
-MODEL_PATH = "/YOUR/MODEL/PATH" # model_download_path: https://huggingface.co/spaces/LiheYoung/Depth-Anything/blob/main/checkpoints/depth_anything_vitb14.pth
+MODEL_PATH = "/YOUR/MODEL/PATH"
 
 ENCODER = "vitl"
 OUTPUT_DIR = str(Path(__file__).parent / ("vis_depth" if DATA_TYPE == "image" else "vis_video_depth"))
-GRAYSCALE = False  # True outputs grayscale image, False outputs color heat map (Only used for image mode)
+GRAYSCALE = False  # # True outputs grayscale image, False outputs color heat map (Only used for image mode)
 
-
-pipeline = DepthAnythingPipeline(
-    encoder=ENCODER,
+pipeline = DepthAnythingPipeline.from_pretrained(
     pretrained_model_path=None if MODEL_PATH == "/YOUR/MODEL/PATH" else MODEL_PATH,
+    encoder=ENCODER,
     data_type=DATA_TYPE,
 )
 
@@ -27,4 +26,3 @@ pipeline(
     outdir=OUTPUT_DIR,
     grayscale=GRAYSCALE,
 )
-
